Moved InitializationEvents to the agent
