+++ android-agent/src/main/java/io/opentelemetry/android/OpenTelemetryRum.java
@@ -73,7 +73,7 @@ public interface OpenTelemetryRum {
      * Get a handle to the instance of the {@linkplain OpenTelemetry OpenTelemetry API} that this
      * instance is using for instrumentation.
      */
-    OpenTelemetrySdk getOpenTelemetry();
+    OpenTelemetry getOpenTelemetry();
 
     /**
      * Get the client session ID associated with this instance of the RUM instrumentation library.
